+++ FacebookAudienceNetwork/src/main/java/com/mopub/mobileads/FacebookInterstitial.java
@@ -17,6 +17,15 @@ import com.mopub.common.logging.MoPubLog;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicBoolean;
 
+import static com.facebook.ads.AdError.BROKEN_MEDIA_ERROR_CODE;
+import static com.facebook.ads.AdError.CACHE_ERROR_CODE;
+import static com.facebook.ads.AdError.INTERNAL_ERROR_CODE;
+import static com.facebook.ads.AdError.INTERSTITIAL_AD_TIMEOUT;
+import static com.facebook.ads.AdError.LOAD_TOO_FREQUENTLY_ERROR_CODE;
+import static com.facebook.ads.AdError.MEDIATION_ERROR_CODE;
+import static com.facebook.ads.AdError.NETWORK_ERROR_CODE;
+import static com.facebook.ads.AdError.NO_FILL_ERROR_CODE;
+import static com.facebook.ads.AdError.SERVER_ERROR_CODE;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CLICKED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CUSTOM;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_ATTEMPTED;
@@ -25,7 +34,15 @@ import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_SUCCESS;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_ATTEMPTED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_FAILED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_SUCCESS;
+import static com.mopub.mobileads.MoPubErrorCode.CANCELLED;
 import static com.mopub.mobileads.MoPubErrorCode.EXPIRED;
+import static com.mopub.mobileads.MoPubErrorCode.NETWORK_INVALID_STATE;
+import static com.mopub.mobileads.MoPubErrorCode.NETWORK_NO_FILL;
+import static com.mopub.mobileads.MoPubErrorCode.NETWORK_TIMEOUT;
+import static com.mopub.mobileads.MoPubErrorCode.NO_CONNECTION;
+import static com.mopub.mobileads.MoPubErrorCode.UNSPECIFIED;
+import static com.mopub.mobileads.MoPubErrorCode.VIDEO_CACHE_ERROR;
+import static com.mopub.mobileads.MoPubErrorCode.VIDEO_PLAYBACK_ERROR;
 
 public class FacebookInterstitial extends CustomEventInterstitial implements InterstitialAdExtendedListener {
     private static final int ONE_HOURS_MILLIS = 60 * 60 * 1000;
@@ -39,6 +56,7 @@ public class FacebookInterstitial extends CustomEventInterstitial implements Int
     private Runnable mAdExpiration;
     @NonNull
     private FacebookAdapterConfiguration mFacebookAdapterConfiguration;
+    private static String mPlacementId;
 
     public FacebookInterstitial() {
         mHandler = new Handler();
@@ -48,11 +66,14 @@ public class FacebookInterstitial extends CustomEventInterstitial implements Int
             @Override
             public void run() {
                 if (mInterstitialListener != null) {
-                    MoPubLog.log(CUSTOM, ADAPTER_NAME, "Expiring unused Facebook Interstitial ad due to Facebook's 60-minute expiration policy.");
+                    MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Expiring unused " +
+                            "Facebook Interstitial ad due to Facebook's 60-minute expiration policy.");
                     mInterstitialListener.onInterstitialFailed(EXPIRED);
-                    MoPubLog.log(LOAD_FAILED, ADAPTER_NAME, MoPubErrorCode.EXPIRED.getIntCode(), MoPubErrorCode.EXPIRED);
+                    MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME,
+                            MoPubErrorCode.EXPIRED.getIntCode(), MoPubErrorCode.EXPIRED);
 
-                    /* Can't get a direct handle to adFailed() to set the interstitial's state to IDLE: https://github.com/mopub/mopub-android-sdk/blob/4199080a1efd755641369715a4de5031d6072fbc/mopub-sdk/mopub-sdk-interstitial/src/main/java/com/mopub/mobileads/MoPubInterstitial.java#L91.
+                    /* Can't get a direct handle to adFailed() to set the interstitial's state to
+                    IDLE: https://github.com/mopub/mopub-android-sdk/blob/4199080a1efd755641369715a4de5031d6072fbc/mopub-sdk/mopub-sdk-interstitial/src/main/java/com/mopub/mobileads/MoPubInterstitial.java#L91.
                     So, invalidating the interstitial (destroying & nulling) instead. */
                     onInvalidate();
                 }
@@ -77,49 +98,51 @@ public class FacebookInterstitial extends CustomEventInterstitial implements Int
 
         mInterstitialListener = customEventInterstitialListener;
 
-        final String placementId;
         if (extrasAreValid(serverExtras)) {
-            placementId = serverExtras.get(PLACEMENT_ID_KEY);
+            mPlacementId = serverExtras.get(PLACEMENT_ID_KEY);
             mFacebookAdapterConfiguration.setCachedInitializationParameters(context, serverExtras);
         } else {
             if (mInterstitialListener != null) {
                 mInterstitialListener.onInterstitialFailed(MoPubErrorCode.NETWORK_NO_FILL);
-                MoPubLog.log(LOAD_FAILED, ADAPTER_NAME, MoPubErrorCode.NETWORK_NO_FILL.getIntCode(), MoPubErrorCode.NETWORK_NO_FILL);
+                MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME,
+                        MoPubErrorCode.NETWORK_NO_FILL.getIntCode(), MoPubErrorCode.NETWORK_NO_FILL);
             }
             return;
         }
 
-        mFacebookInterstitial = new InterstitialAd(context, placementId);
+        mFacebookInterstitial = new InterstitialAd(context, mPlacementId);
 
         final String adm = serverExtras.get(DataKeys.ADM_KEY);
 
-        InterstitialAd.InterstitialAdLoadConfigBuilder interstitialLoadAdConfigBuilder = mFacebookInterstitial
-                .buildLoadAdConfig()
-                .withAdListener(this);
+        InterstitialAd.InterstitialAdLoadConfigBuilder interstitialLoadAdConfigBuilder =
+                mFacebookInterstitial.buildLoadAdConfig().withAdListener(this);
 
         if (!TextUtils.isEmpty(adm)) {
             mFacebookInterstitial.loadAd(interstitialLoadAdConfigBuilder.withBid(adm).build());
-            MoPubLog.log(placementId, LOAD_ATTEMPTED, ADAPTER_NAME);
+            MoPubLog.log(getAdNetworkId(), LOAD_ATTEMPTED, ADAPTER_NAME);
         } else {
             mFacebookInterstitial.loadAd(interstitialLoadAdConfigBuilder.build());
-            MoPubLog.log(placementId, LOAD_ATTEMPTED, ADAPTER_NAME);
+            MoPubLog.log(getAdNetworkId(), LOAD_ATTEMPTED, ADAPTER_NAME);
         }
     }
 
     @Override
     protected void showInterstitial() {
-        MoPubLog.log(SHOW_ATTEMPTED, ADAPTER_NAME);
+        MoPubLog.log(getAdNetworkId(), SHOW_ATTEMPTED, ADAPTER_NAME);
         if (mFacebookInterstitial != null && mFacebookInterstitial.isAdLoaded() &&
                 !mFacebookInterstitial.isAdInvalidated()) {
             mFacebookInterstitial.show();
             cancelExpirationTimer();
         } else {
-            MoPubLog.log(SHOW_FAILED, ADAPTER_NAME, MoPubErrorCode.NETWORK_NO_FILL.getIntCode(), MoPubErrorCode.NETWORK_NO_FILL);
-            MoPubLog.log(CUSTOM, ADAPTER_NAME, "Tried to show a Facebook interstitial ad when it's not ready. Please try again.");
+            MoPubLog.log(getAdNetworkId(), SHOW_FAILED, ADAPTER_NAME,
+                    MoPubErrorCode.NETWORK_NO_FILL.getIntCode(), MoPubErrorCode.NETWORK_NO_FILL);
+            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Tried to show a Facebook " +
+                    "interstitial ad when it's not ready. Please try again.");
             if (mInterstitialListener != null) {
                 onError(mFacebookInterstitial, AdError.INTERNAL_ERROR);
             } else {
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "Interstitial listener not instantiated. Please load interstitial again.");
+                MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Interstitial listener " +
+                        "not instantiated. Please load interstitial again.");
             }
         }
     }
@@ -143,7 +166,7 @@ public class FacebookInterstitial extends CustomEventInterstitial implements Int
         cancelExpirationTimer();
         if (mInterstitialListener != null) {
             mInterstitialListener.onInterstitialLoaded();
-            MoPubLog.log(LOAD_SUCCESS, ADAPTER_NAME);
+            MoPubLog.log(getAdNetworkId(), LOAD_SUCCESS, ADAPTER_NAME);
         }
         mHandler.postDelayed(mAdExpiration, ONE_HOURS_MILLIS);
     }
@@ -151,24 +174,51 @@ public class FacebookInterstitial extends CustomEventInterstitial implements Int
     @Override
     public void onError(final Ad ad, final AdError error) {
         cancelExpirationTimer();
+        MoPubErrorCode errorCode;
+
+        switch (error.getErrorCode()) {
+            case NO_FILL_ERROR_CODE:
+                errorCode = NETWORK_NO_FILL;
+                break;
+            case INTERNAL_ERROR_CODE:
+                errorCode = MoPubErrorCode.INTERNAL_ERROR;
+                break;
+            case NETWORK_ERROR_CODE:
+                errorCode = NO_CONNECTION;
+                break;
+            case LOAD_TOO_FREQUENTLY_ERROR_CODE:
+                errorCode = CANCELLED;
+                break;
+            case SERVER_ERROR_CODE:
+                errorCode = MoPubErrorCode.SERVER_ERROR;
+                break;
+            case CACHE_ERROR_CODE:
+                errorCode = VIDEO_CACHE_ERROR;
+                break;
+            case MEDIATION_ERROR_CODE:
+                errorCode = NETWORK_INVALID_STATE;
+                break;
+            case INTERSTITIAL_AD_TIMEOUT:
+                errorCode = NETWORK_TIMEOUT;
+                break;
+            case BROKEN_MEDIA_ERROR_CODE:
+                errorCode = VIDEO_PLAYBACK_ERROR;
+                break;
+            default:
+                errorCode = UNSPECIFIED;
+        }
+
+        MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME, errorCode.getIntCode(), errorCode);
+
         if (mInterstitialListener != null) {
-            if (error.getErrorCode() == AdError.NO_FILL.getErrorCode()) {
-                mInterstitialListener.onInterstitialFailed(MoPubErrorCode.NETWORK_NO_FILL);
-                MoPubLog.log(LOAD_FAILED, ADAPTER_NAME, MoPubErrorCode.NETWORK_NO_FILL.getIntCode(), MoPubErrorCode.NETWORK_NO_FILL);
-            } else if (error.getErrorCode() == AdError.INTERNAL_ERROR.getErrorCode()) {
-                mInterstitialListener.onInterstitialFailed(MoPubErrorCode.NETWORK_INVALID_STATE);
-                MoPubLog.log(LOAD_FAILED, ADAPTER_NAME, MoPubErrorCode.NETWORK_INVALID_STATE.getIntCode(), MoPubErrorCode.NETWORK_INVALID_STATE);
-            } else {
-                mInterstitialListener.onInterstitialFailed(MoPubErrorCode.UNSPECIFIED);
-                MoPubLog.log(LOAD_FAILED, ADAPTER_NAME, MoPubErrorCode.UNSPECIFIED.getIntCode(), MoPubErrorCode.UNSPECIFIED);
-            }
+            mInterstitialListener.onInterstitialFailed(errorCode);
         }
     }
 
     @Override
     public void onInterstitialDisplayed(final Ad ad) {
         cancelExpirationTimer();
-        MoPubLog.log(SHOW_SUCCESS, ADAPTER_NAME);
+        MoPubLog.log(getAdNetworkId(), SHOW_SUCCESS, ADAPTER_NAME);
         if (mInterstitialListener != null) {
             mInterstitialListener.onInterstitialShown();
         }
@@ -176,7 +226,7 @@ public class FacebookInterstitial extends CustomEventInterstitial implements Int
 
     @Override
     public void onAdClicked(final Ad ad) {
-        MoPubLog.log(CLICKED, ADAPTER_NAME);
+        MoPubLog.log(getAdNetworkId(), CLICKED, ADAPTER_NAME);
         if (mInterstitialListener != null) {
             mInterstitialListener.onInterstitialClicked();
         }
@@ -184,7 +234,8 @@ public class FacebookInterstitial extends CustomEventInterstitial implements Int
 
     @Override
     public void onLoggingImpression(Ad ad) {
-        MoPubLog.log(CUSTOM, ADAPTER_NAME, "Facebook interstitial ad logged impression.");
+        MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Facebook interstitial ad " +
+                "logged impression.");
         if (mInterstitialListener != null) {
             mInterstitialListener.onInterstitialImpression();
         }
@@ -226,5 +277,9 @@ public class FacebookInterstitial extends CustomEventInterstitial implements Int
 
     private void cancelExpirationTimer() {
         mHandler.removeCallbacks(mAdExpiration);
+    }
+
+    private static String getAdNetworkId() {
+        return mPlacementId;
     }
 }