+++ FacebookAudienceNetwork/src/main/java/com/mopub/nativeads/FacebookNative.java
@@ -27,19 +27,34 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicBoolean;
 
+import static com.facebook.ads.AdError.CACHE_ERROR_CODE;
+import static com.facebook.ads.AdError.LOAD_TOO_FREQUENTLY_ERROR_CODE;
+import static com.facebook.ads.AdError.MEDIATION_ERROR_CODE;
+import static com.facebook.ads.AdError.NETWORK_ERROR_CODE;
+import static com.facebook.ads.AdError.NO_FILL_ERROR_CODE;
+import static com.facebook.ads.AdError.SERVER_ERROR_CODE;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CLICKED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CUSTOM;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_ATTEMPTED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_FAILED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_SUCCESS;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_SUCCESS;
+import static com.mopub.nativeads.NativeErrorCode.CONNECTION_ERROR;
+import static com.mopub.nativeads.NativeErrorCode.IMAGE_DOWNLOAD_FAILURE;
+import static com.mopub.nativeads.NativeErrorCode.INVALID_RESPONSE;
+import static com.mopub.nativeads.NativeErrorCode.NETWORK_INVALID_REQUEST;
+import static com.mopub.nativeads.NativeErrorCode.NETWORK_NO_FILL;
+import static com.mopub.nativeads.NativeErrorCode.UNEXPECTED_RESPONSE_CODE;
+import static com.mopub.nativeads.NativeErrorCode.UNSPECIFIED;
 
 public class FacebookNative extends CustomEventNative {
+    public static final String NATIVE_BANNER_KEY = "native_banner";
+
     private static final String PLACEMENT_ID_KEY = "placement_id";
-    private static final String NATIVE_BANNER_KEY = "native_banner";
     private static final String ADAPTER_NAME = FacebookNative.class.getSimpleName();
     private static AtomicBoolean sIsInitialized = new AtomicBoolean(false);
     private Boolean isNativeBanner;
+    private static String mPlacementId;
 
     @NonNull
     private FacebookAdapterConfiguration mFacebookAdapterConfiguration;
@@ -54,6 +69,10 @@ public class FacebookNative extends CustomEventNative {
                                 @NonNull final CustomEventNativeListener customEventNativeListener,
                                 @NonNull final Map<String, Object> localExtras,
                                 @NonNull final Map<String, String> serverExtras) {
+        Preconditions.checkNotNull(context);
+        Preconditions.checkNotNull(customEventNativeListener);
+        Preconditions.checkNotNull(localExtras);
+        Preconditions.checkNotNull(serverExtras);
 
         if (!sIsInitialized.getAndSet(true)) {
             AudienceNetworkAds.initialize(context);
@@ -63,8 +82,9 @@ public class FacebookNative extends CustomEventNative {
             placementId = serverExtras.get(PLACEMENT_ID_KEY);
             mFacebookAdapterConfiguration.setCachedInitializationParameters(context, serverExtras);
         } else {
-            customEventNativeListener.onNativeAdFailed(NativeErrorCode.NETWORK_NO_FILL);
-            MoPubLog.log(LOAD_FAILED, ADAPTER_NAME, NativeErrorCode.NETWORK_NO_FILL.getIntCode(), NativeErrorCode.NETWORK_NO_FILL);
+            customEventNativeListener.onNativeAdFailed(NETWORK_NO_FILL);
+            MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME,
+                    NETWORK_NO_FILL.getIntCode(), NETWORK_NO_FILL);
             return;
         }
 
@@ -79,13 +99,14 @@ public class FacebookNative extends CustomEventNative {
 
         // The native banner flag is not set in localExtras.
         // Fall back to retrieving the flag from the FacebookAdapterConfiguration.
-        isNativeBanner = isNativeBanner == null ? FacebookAdapterConfiguration.getNativeBannerPref() : isNativeBanner;
+        isNativeBanner = isNativeBanner == null ? FacebookAdapterConfiguration.getNativeBannerPref()
+                : isNativeBanner;
 
         if (isNativeBanner != null) {
             if (isNativeBanner) {
                 final FacebookNativeAd facebookNativeBannerAd =
-                        new FacebookNativeAd(context,
-                                new NativeBannerAd(context, placementId), customEventNativeListener, bid);
+                        new FacebookNativeAd(context, new NativeBannerAd(context, placementId),
+                                customEventNativeListener, bid);
                 facebookNativeBannerAd.loadAd();
 
                 return;
@@ -101,12 +122,14 @@ public class FacebookNative extends CustomEventNative {
     }
 
     private boolean extrasAreValid(final Map<String, String> serverExtras) {
-        final String placementId = serverExtras.get(PLACEMENT_ID_KEY);
-        return (!TextUtils.isEmpty(placementId));
+        mPlacementId = serverExtras.get(PLACEMENT_ID_KEY);
+        return (!TextUtils.isEmpty(mPlacementId));
     }
 
-    private static void registerChildViewsForInteraction(final View view, final NativeAdBase nativeAdBase,
-                                                         @Nullable final MediaView mediaView, final MediaView adIconView) {
+    private static void registerChildViewsForInteraction(final View view,
+                                                         final NativeAdBase nativeAdBase,
+                                                         @Nullable final MediaView mediaView,
+                                                         final MediaView adIconView) {
 
         if (nativeAdBase == null) {
             return;
@@ -135,12 +158,13 @@ public class FacebookNative extends CustomEventNative {
     private static void assembleChildViewsWithLimit(final View view,
                                                     final List<View> clickableViews, final int limit) {
         if (view == null) {
-            MoPubLog.log(CUSTOM, "View given is null. Ignoring");
+            MoPubLog.log(getAdNetworkId(), CUSTOM, "View given is null. Ignoring");
             return;
         }
 
         if (limit <= 0) {
-            MoPubLog.log(CUSTOM, "Depth limit reached; adding this view regardless of its type.");
+            MoPubLog.log(getAdNetworkId(), CUSTOM, "Depth limit reached; adding this view " +
+                    "regardless of its type.");
             clickableViews.add(view);
             return;
         }
@@ -156,10 +180,13 @@ public class FacebookNative extends CustomEventNative {
         clickableViews.add(view);
     }
 
+    private static String getAdNetworkId() {
+        return mPlacementId;
+    }
+
     static class FacebookNativeAd extends BaseNativeAd implements NativeAdListener {
         private static final String SOCIAL_CONTEXT_FOR_AD = "socialContextForAd";
 
-        private final Context mContext;
         private final NativeAdBase mNativeAd;
         private final CustomEventNativeListener mCustomEventNativeListener;
 
@@ -171,7 +198,6 @@ public class FacebookNative extends CustomEventNative {
                          final NativeAdBase nativeAd,
                          final CustomEventNativeListener customEventNativeListener,
                          final String bid) {
-            mContext = context.getApplicationContext();
             mNativeAd = nativeAd;
             mCustomEventNativeListener = customEventNativeListener;
             mExtras = new HashMap<String, Object>();
@@ -185,10 +211,10 @@ public class FacebookNative extends CustomEventNative {
 
             if (!TextUtils.isEmpty(mBid)) {
                 mNativeAd.loadAd(nativeAdLoadConfigBuilder.withBid(mBid).build());
-                MoPubLog.log(mBid, LOAD_ATTEMPTED, ADAPTER_NAME);
+                MoPubLog.log(getAdNetworkId(), LOAD_ATTEMPTED, ADAPTER_NAME);
             } else {
                 mNativeAd.loadAd(nativeAdLoadConfigBuilder.build());
-                MoPubLog.log(LOAD_ATTEMPTED, ADAPTER_NAME);
+                MoPubLog.log(getAdNetworkId(), LOAD_ATTEMPTED, ADAPTER_NAME);
             }
         }
 
@@ -244,52 +270,73 @@ public class FacebookNative extends CustomEventNative {
             // This identity check is from Facebook's Native API sample code:
             // https://developers.facebook.com/docs/audience-network/android/native-api
             if (!mNativeAd.equals(ad) || !mNativeAd.isAdLoaded()) {
-                mCustomEventNativeListener.onNativeAdFailed(NativeErrorCode.NETWORK_NO_FILL);
-                MoPubLog.log(LOAD_FAILED, ADAPTER_NAME, NativeErrorCode.NETWORK_NO_FILL.getIntCode(), NativeErrorCode.NETWORK_NO_FILL);
+                mCustomEventNativeListener.onNativeAdFailed(NETWORK_NO_FILL);
+                MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME,
+                        NETWORK_NO_FILL.getIntCode(), NETWORK_NO_FILL);
                 return;
             }
 
             addExtra(SOCIAL_CONTEXT_FOR_AD, mNativeAd.getAdSocialContext());
             mCustomEventNativeListener.onNativeAdLoaded(FacebookNativeAd.this);
-            MoPubLog.log(LOAD_SUCCESS, ADAPTER_NAME);
+            MoPubLog.log(getAdNetworkId(), LOAD_SUCCESS, ADAPTER_NAME);
         }
 
         @Override
         public void onError(final Ad ad, final AdError adError) {
-            if (adError == null) {
-                mCustomEventNativeListener.onNativeAdFailed(NativeErrorCode.UNSPECIFIED);
-                MoPubLog.log(LOAD_FAILED, ADAPTER_NAME, NativeErrorCode.UNSPECIFIED.getIntCode(), NativeErrorCode.UNSPECIFIED);
-            } else if (adError.getErrorCode() == AdError.NO_FILL.getErrorCode()) {
-                mCustomEventNativeListener.onNativeAdFailed(NativeErrorCode.NETWORK_NO_FILL);
-                MoPubLog.log(LOAD_FAILED, ADAPTER_NAME, NativeErrorCode.NETWORK_NO_FILL.getIntCode(), NativeErrorCode.NETWORK_NO_FILL);
-            } else if (adError.getErrorCode() == AdError.INTERNAL_ERROR.getErrorCode()) {
-                mCustomEventNativeListener.onNativeAdFailed(NativeErrorCode.NETWORK_INVALID_STATE);
-                MoPubLog.log(LOAD_FAILED, ADAPTER_NAME, NativeErrorCode.NETWORK_INVALID_STATE.getIntCode(), NativeErrorCode.NETWORK_INVALID_STATE);
-            } else {
-                mCustomEventNativeListener.onNativeAdFailed(NativeErrorCode.UNSPECIFIED);
-                MoPubLog.log(LOAD_FAILED, ADAPTER_NAME, NativeErrorCode.UNSPECIFIED.getIntCode(), NativeErrorCode.UNSPECIFIED);
+            NativeErrorCode errorCode;
+
+            switch (adError.getErrorCode()) {
+                case NO_FILL_ERROR_CODE:
+                    errorCode = NETWORK_NO_FILL;
+                    break;
+                case NETWORK_ERROR_CODE:
+                    errorCode = CONNECTION_ERROR;
+                    break;
+                case LOAD_TOO_FREQUENTLY_ERROR_CODE:
+                    errorCode = NETWORK_INVALID_REQUEST;
+                    break;
+                case SERVER_ERROR_CODE:
+                    errorCode = INVALID_RESPONSE;
+                    break;
+                case CACHE_ERROR_CODE:
+                    errorCode = IMAGE_DOWNLOAD_FAILURE;
+                    break;
+                case MEDIATION_ERROR_CODE:
+                    errorCode = UNEXPECTED_RESPONSE_CODE;
+                    break;
+                default:
+                    errorCode = UNSPECIFIED;
             }
+
+            MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME, errorCode.getIntCode(),
+                    errorCode);
+
+            if (mCustomEventNativeListener != null) {
+                mCustomEventNativeListener.onNativeAdFailed(errorCode);
+            }
+
         }
 
         @Override
         public void onAdClicked(final Ad ad) {
             notifyAdClicked();
-            MoPubLog.log(CLICKED, ADAPTER_NAME);
+            MoPubLog.log(getAdNetworkId(), CLICKED, ADAPTER_NAME);
         }
 
         @Override
         public void onLoggingImpression(final Ad ad) {
             notifyAdImpressed();
-            MoPubLog.log(SHOW_SUCCESS, ADAPTER_NAME);
+            MoPubLog.log(getAdNetworkId(), SHOW_SUCCESS, ADAPTER_NAME);
         }
 
         // BaseForwardingNativeAd
         @Override
-        public void prepare(final View view) {
+        public void prepare(@NonNull View view) {
         }
 
         @Override
-        public void clear(final View view) {
+        public void clear(@NonNull View view) {
+            Preconditions.checkNotNull(view);
             mNativeAd.unregisterView();
         }
 
@@ -304,7 +351,8 @@ public class FacebookNative extends CustomEventNative {
          */
         @Nullable
         final public Object getExtra(final String key) {
-            if (!Preconditions.NoThrow.checkNotNull(key, "getExtra key is not allowed to be null")) {
+            if (!Preconditions.NoThrow.checkNotNull(key, "getExtra key is not allowed " +
+                    "to be null")) {
                 return null;
             }
             return mExtras.get(key);
@@ -321,7 +369,8 @@ public class FacebookNative extends CustomEventNative {
         }
 
         final public void addExtra(final String key, final Object value) {
-            if (!Preconditions.NoThrow.checkNotNull(key, "addExtra key is not allowed to be null")) {
+            if (!Preconditions.NoThrow.checkNotNull(key, "addExtra key is not allowed " +
+                    "to be null")) {
                 return;
             }
             mExtras.put(key, value);