+++ FacebookAudienceNetwork/src/main/java/com/mopub/nativeads/FacebookTemplateRenderer.java
@@ -28,6 +28,7 @@ public class FacebookTemplateRenderer implements MoPubAdRenderer<FacebookNative.
     @NonNull
     @Override
     public View createAdView(@NonNull Context context, @Nullable ViewGroup parent) {
+        Preconditions.checkNotNull(context);
         return new FrameLayout(context);
     }
 
@@ -39,9 +40,11 @@ public class FacebookTemplateRenderer implements MoPubAdRenderer<FacebookNative.
         NativeAdBase nativeAdBase = ad.getFacebookNativeAd();
 
         if (nativeAdBase instanceof NativeAd) {
-            View adView = NativeAdView.render(parentView.getContext(), (NativeAd) nativeAdBase, mTemplateAttributes);
+            View adView = NativeAdView.render(parentView.getContext(), (NativeAd) nativeAdBase,
+                    mTemplateAttributes);
 
-            FrameLayout.LayoutParams adViewParams = new FrameLayout.LayoutParams(WRAP_CONTENT, WRAP_CONTENT);
+            FrameLayout.LayoutParams adViewParams = new FrameLayout.LayoutParams(WRAP_CONTENT,
+                    WRAP_CONTENT);
             ((FrameLayout) parentView).addView(adView, adViewParams);
         }
     }