+++ FacebookAudienceNetwork/CHANGELOG.md
@@ -1,4 +1,9 @@
 ## Changelog
+ * 5.6.0.1
+    * Log the Facebook placement name in ad lifecycle events.
+    * Map additional error codes for failure cases.
+    * Fail fast when certain parameters are null. 
+
  * 5.6.0.0
     * This version of the adapters has been certified with Facebook Audience Network 5.6.0.
     * Add support for [native banner](https://developers.facebook.com/docs/audience-network/android-native-banner/). Refer to the [Mediate Facebook page](https://developers.mopub.com/publishers/mediation/networks/facebook/) for integration instructions.